Topology issue: importing unnecessary modules in a lambda inside pypi package function
Topology is generating code load the streamsx.xyz pypi package inside a
map(lambda t:t)  Python operator even when nothing from the module is
needed. This is somehow understandable because a lambda may use
functions from module and it has no parameter to determine needed
imports.

Will check if we can include the package even when we don't need it in
the running application.

But for now the code is running having the map(lambda t:t) outside the
streamsx.xyz package defined function (in the notebook)